배틀 - 매칭 연동 테스트 중

+++ MasterServer/MasterServer/LanServer.h
@@ -113,7 +113,7 @@ typedef struct st_BattleServer
 
 	WCHAR	ChatServerIP[16];
 	WORD	ChatServerPort;
-	long	ServerNo;
+	WORD	ServerNo;
 }BattleServer;
 
 #endif
